+++ src/diff.ts
@@ -1,10 +1,9 @@
-// Version:3.1.0
+// Version:3.2.0
 
 import { state, useTemplate } from './init.js';
 import {
 	getType,
 	checkVnode,
-	isHasFlag,
 	isComplexType,
 	isXlink,
 	isSameObject,
@@ -28,25 +27,79 @@ export interface vnodeType {
 	el?: HTMLElement;
 }
 
-const _com_ = Object.create(null);
-const _components = new WeakMap();
-const flag = ['$key', '$name'];
-let componentName = '';
+const _com_: any = Object.create(null);
+const _components: WeakMap<object, any> = new WeakMap();
+const flag: Array<string> = ['$key', '$name', '$props'];
+let componentName: string = '';
+export const domInfo: any = Object.create(null);
+export let propsData: any = reactive(Object.create(null));
 
-function mount(vnode: vnodeType, container: HTMLElement | Node, anchor?: Node) {
+interface proxyConfType {
+	[propName: string]: any;
+}
+
+function reactive<T extends proxyConfType>(target: object = {}): any {
+	if (typeof target !== 'object' || target === null) {
+		return target;
+	}
+
+	const proxyConf = {
+		get(target: T, key: string, receiver: any): any {
+			const ownKeys = Reflect.ownKeys(target);
+			if (ownKeys.includes(key)) {
+				const result = Reflect.get(target, key, receiver);
+				return reactive(result);
+			}
+		},
+		set(target: T, key: string, val: any, receiver: any): boolean {
+			if (val === target[key]) {
+				return true;
+			}
+
+			const ownKeys = Reflect.ownKeys(target);
+			if (ownKeys.includes(key) || Object.keys(_com_).includes(key)) {
+				const result = Reflect.set(target, key, val, receiver);
+				return result;
+			}
+		},
+		deleteProperty(target: T, key: string): boolean {
+			const result = Reflect.deleteProperty(target, key);
+			return result;
+		},
+	};
+
+	const observed = new Proxy(target, proxyConf);
+	return observed;
+}
+
+function mount(
+	vnode: vnodeType,
+	container: HTMLElement | Node,
+	anchor?: Node
+): void {
 	if (vnode.tag) {
 		const el: any = createNode(vnode.tag);
 
 		if (vnode.props) {
 			addEvent(el, vnode.props);
 
-			if (isHasFlag(vnode.props, flag[0])) {
+			if (vnode.props.hasOwnProperty(flag[0])) {
 				vnode.el = el;
+				if (getType(vnode.props[flag[0]]) === 'string') {
+					domInfo[vnode.props[flag[0]]] = el;
+				}
 			}
 
-			if (vnode.props[flag[1]]) {
+			if (vnode.props.hasOwnProperty(flag[1])) {
 				_com_[vnode.props[flag[1]]] = Object.create(null);
 				_components.set(_com_[vnode.props[flag[1]]], vnode.children[0]);
+			}
+
+			if (
+				vnode.props.hasOwnProperty(flag[1]) &&
+				vnode.props.hasOwnProperty(flag[2])
+			) {
+				propsData[vnode.props[flag[1]]] = vnode.props[flag[2]];
 			}
 
 			for (const key in vnode.props) {
@@ -55,7 +108,7 @@ function mount(vnode: vnodeType, container: HTMLElement | Node, anchor?: Node) {
 						if (isXlink(key)) {
 							el.setAttributeNS(xlinkNS, key, vnode.props[key]);
 						} else {
-							if (key !== flag[0] && key !== flag[1]) {
+							if (!flag.includes(key)) {
 								el.setAttribute(key, vnode.props[key]);
 							}
 						}
@@ -96,17 +149,17 @@ function mount(vnode: vnodeType, container: HTMLElement | Node, anchor?: Node) {
 	}
 }
 
-function patch(n1: vnodeType, n2: vnodeType, status: string) {
-	const oldProps = n1.props || {};
+function patch(n1: vnodeType, n2: vnodeType, status: string): void {
+	const oldProps: any = n1.props || {};
 
-	if (isHasFlag(oldProps, flag[0]) && n1.tag !== n2.tag) {
+	if (oldProps.hasOwnProperty(flag[0]) && n1.tag !== n2.tag) {
 		const parent = n1.el.parentNode;
 		const anchor = n1.el.nextSibling;
 		parent.removeChild(n1.el);
 		mount(n2, parent, anchor);
 	} else {
-		let el: null | HTMLElement | any = null;
-		if (isHasFlag(oldProps, flag[0])) {
+		let el: any = null;
+		if (oldProps.hasOwnProperty(flag[0])) {
 			const newProps = n2.props || {};
 			el = n2.el = n1.el;
 
@@ -114,7 +167,7 @@ function patch(n1: vnodeType, n2: vnodeType, status: string) {
 				let [newValue, oldValue] = [newProps[key], oldProps[key]];
 				if (newValue !== oldValue) {
 					if (newValue !== null) {
-						if (getType(newValue) !== 'function' && key !== flag[0]) {
+						if (getType(newValue) !== 'function' && !flag.includes(key)) {
 							el[key] && (el[key] = newValue); // property
 
 							if (isXlink(key)) {
@@ -147,10 +200,10 @@ function patch(n1: vnodeType, n2: vnodeType, status: string) {
 			}
 		}
 
-		const oc = n1.children[0];
-		const nc = n2.children[0];
-		const ocs = n1.children;
-		const ncs = n2.children;
+		const oc: any = n1.children[0];
+		const nc: any = n2.children[0];
+		const ocs: any[] = n1.children;
+		const ncs: any[] = n2.children;
 
 		if (!isSameObject(ocs, ncs)) {
 			updateChildrenNode(ncs, el, patchChildren);
@@ -169,7 +222,7 @@ function patch(n1: vnodeType, n2: vnodeType, status: string) {
 	}
 }
 
-function patchNode(o: any[], n: any[], el: HTMLElement, status: string) {
+function patchNode(o: any[], n: any[], el: HTMLElement, status: string): void {
 	if (status === 'useFirstKey') {
 		for (let i = 1; i <= Math.max(o.length, n.length); i++) {
 			if (!o[o.length - i]) {
@@ -198,7 +251,7 @@ function updateChildrenNode(
 	childNode: any[],
 	el: HTMLElement,
 	setChildrenNode: Function
-) {
+): void {
 	if (childNode.length === 1 && !isComplexType(childNode[0])) {
 		el && updateTextNode(childNode, el);
 	} else if (childNode.length > 1 && !checkVnode(childNode)) {
@@ -214,7 +267,7 @@ function updateChildrenNode(
 	}
 }
 
-function updateTextNode(val: any, el: HTMLElement) {
+function updateTextNode(val: any, el: HTMLElement): void {
 	if (isComplexType(val)) {
 		if (
 			getType(val) === 'function' ||
@@ -230,19 +283,35 @@ function updateTextNode(val: any, el: HTMLElement) {
 	}
 }
 
+let mountHook: Function | null = null;
+export function onMounted(fn: Function): void {
+	mountHook = fn;
+}
+
+let unMountedHook: Function | null = null;
+export function onUnmounted(fn: Function): void {
+	unMountedHook = fn;
+}
+
+let nextTickHook: Function | null = null;
+export function nextTick(fn: Function): void {
+	nextTickHook = fn;
+}
+
 export function mountNode(
 	dom: vnodeType,
 	selector: HTMLElement,
 	status?: string,
 	name?: string
-) {
+): void {
 	if (!state.isMounted) {
-		const _template = useFragmentNode(dom);
+		const _template: vnodeType = useFragmentNode(dom);
 		mount(_template, selector);
 		state.oldTree = _template;
 		state.isMounted = true;
+		mountHook && mountHook();
 	} else {
-		const newTree = useFragmentNode(dom);
+		const newTree: vnodeType = useFragmentNode(dom);
 		patch(state.oldTree, newTree, status);
 		state.oldTree = newTree;
 
@@ -252,7 +321,10 @@ export function mountNode(
 	}
 }
 
-export function setData(callback: Function, options: setDataOptionsType) {
+export function setData(
+	callback: Function,
+	options: setDataOptionsType
+): Promise<void> {
 	if (getType(callback) === 'function' && getType(Promise) !== 'undefined') {
 		return Promise.resolve()
 			.then(() => {
@@ -260,11 +332,14 @@ export function setData(callback: Function, options: setDataOptionsType) {
 			})
 			.then(() => {
 				if (options && options.status === 'useRouter') {
+					unMountedHook && unMountedHook();
 					state._el.innerHTML = '';
+					unMountedHook = null;
 					mount((state.oldTree = useTemplate(state._template())), state._el);
+					mountHook && mountHook();
 				} else if (options && typeof options.name === 'function') {
-					const name = options.name.name;
-					const _component = options.name();
+					const name: string = options.name.name;
+					const _component: vnodeType = options.name();
 
 					if (componentName !== name) {
 						componentName = name;
@@ -273,9 +348,11 @@ export function setData(callback: Function, options: setDataOptionsType) {
 
 					mountNode(_component, state._el, options.status, name);
 				} else {
-					const status = options && options.status ? options.status : null;
+					const status: string | null =
+						options && options.status ? options.status : null;
 					mountNode(useTemplate(state._template()), state._el, status);
 				}
+				nextTickHook && nextTickHook();
 			})
 			.catch((err) => console.error(err));
 	}